Started Project, Setup main function
Today I started the class and the project for the TrippleX game. The class goes over what is the main() function. We also learned how to cout as well as declare and initialize variables. I set up a welcome message for the game via cout. The instructor also taught us about the const keyword.