Finished the Game
I added two new fuctions. The play game function takes a difficulty argument and makes the game harder based on the difficulty level. The game gets more difficult the more levels you complete. I also added a helper function that handles printing the welcome message for each level. This also takes the difficulty parameter and uses it to print out which level the player is on.

The game uses rand() to pick a random range. The range is computed to be the difficulty + the difficulty.

Then the player has to guess what numbers the computer generated. If they get it right they move on to the next level, increasing the difficulty. If not, they have to retry the level.

Once they hit level 10 the game ends.